requirements.txt
streamlit==1.32.2
openai>=1.10.0,<2.0.0
pinecone-client==3.0.2
langchain==0.1.12
langchain-openai==0.0.8
langchain-core>=0.1.31,<0.2.0
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.0 